catalyst_web/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
# Editable Git install with no remote (catalyst-agent==0.1.0)
-e c:\users\rbui\git\ml\catalyst\catalyst_agent
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.69.0
pillow==11.1.0
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
